Remove JSONPath
It's a dead spec, and it's a dead library. Piping to `jq` is a
perfectly cromulent solution, and it's more flexible than using an
internal path selector. If this becomes a problem later on, or if I
feel saucy then I will implement support for github.com/itchyny/gojq
as a replacement.